fixed two errors and relocated the classes
