+++ src/csvcubed/utils/qb/validation/uri_safe.py
@@ -4,22 +4,21 @@ URI Safe Validations
 """
 from typing import Dict, List, Set, Tuple, Type, Union
 
-from csvcubed.models.cube.qb.components.validationerrors import (
-    ConflictingUriSafeValuesError,
-)
+from csvcubed.models.validationerror import ConflictingUriSafeValuesError
 
 
 def ensure_no_uri_safe_conflicts(
     label_with_uri_safe_identifier: List[Tuple[str, str]],
     location: Union[Type, str],
-) -> None:
+    property_path: List[str],
+    offending_value: List,
+) -> List[ConflictingUriSafeValuesError]:
     """
     Accepts a list of tuples of `(label, uri_safe_identifier)` and identifies any duplicate mappings of
     `label` => `uri_safe_identifier`.
 
     :raises ConflictingUriSafeValuesError when conflicts are found.
     """
-
     map_uri_safe_val_to_labels: Dict[str, Set[str]] = {}
 
     for (label, uri_safe_identifier) in label_with_uri_safe_identifier:
@@ -38,4 +37,14 @@ def ensure_no_uri_safe_conflicts(
     }
 
     if any(collisions):
-        raise ConflictingUriSafeValuesError(location, collisions)
+        return [
+            ConflictingUriSafeValuesError(
+                message="",  # this variable will be populated in the post__init__
+                component_type=location,
+                map_uri_safe_values_to_conflicting_labels=collisions,
+                property_path=property_path,
+                offending_value=offending_value,
+            )
+        ]
+
+    return []
