+++ tests/unit/cube/qb/test_cubeqb_errorvalidation.py
@@ -22,10 +22,6 @@ from csvcubed.models.cube.qb.components.measuresdimension import QbMultiMeasureD
 from csvcubed.models.cube.qb.components.observedvalue import QbObservationValue
 from csvcubed.models.cube.qb.components.unit import ExistingQbUnit, NewQbUnit
 from csvcubed.models.cube.qb.components.unitscolumn import QbMultiUnits
-from csvcubed.models.cube.qb.components.validationerrors import (
-    ConflictingUriSafeValuesError,
-    ReservedUriValueError,
-)
 from csvcubed.models.cube.qb.validationerrors import (
     AttributeNotLinkedError,
     BothMeasureTypesDefinedError,
@@ -41,6 +37,10 @@ from csvcubed.models.cube.qb.validationerrors import (
     NoUnitsDefinedError,
     PivotedObsValColWithoutMeasureError,
     PivotedShapeMeasureColumnsExistError,
+)
+from csvcubed.models.validationerror import (
+    ConflictingUriSafeValuesError,
+    ReservedUriValueError,
 )
 from csvcubed.utils.qb.validation.cube import validate_qb_component_constraints
 from tests.unit.test_baseunit import *
